requirements.txt
black>=23.9.1,<24.0
Django>=4.2.5,<4.3
django-debug-toolbar>=4.2.0,<4.3
django-filter>=23.3,<23.4
djangorestframework>=3.14.0,<3.15
flake8>=6.1.0,<6.2
pillow>=9.5.0,<9.6
psycopg>=3.1.9,<3.2